- web ui development
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@2316 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/io/File.cpp
@@ -169,11 +169,13 @@ bool File::canRead() {
 
 bool File::canWrite() {
   bool result = false;
-  FILE* fh = fopen(getPath().c_str(), "a");
-  if (fh) {
-    result = true;
-    fclose(fh);
-    deleteFile();
+  if(!exists()){
+    FILE* fh = fopen(getPath().c_str(), "a");
+    if (fh) {
+      result = true;
+      fclose(fh);
+      deleteFile();
+    }
   }
   return result;
 }
